Get rid of DynDns mode in favor of allowing a user to force to use static host/port in the settings dialog.
That now takes precedence, if enabled. If not enabled, and UPnP is enabled, it will attempt UPnP, falling back to static host/port if UPnP doesn't work, falling back to LAN if not.

Users can still set LAN specifically in the config file, if they wish. Maybe should expose that at some point.