Add twitter info plugin. Still needs some work -- when it should delete itself (for instance if you disable the plugin and re-enable) it doesn't seem to, leading to twitter complaining about duplicate statuses. Also get a message about creating children for parent in different thread.
Along the way, made lastfm plugin correctly switch to the right thread
for the info plugin, fix a couple bugs (such as loving and unloving
using the same type), and so on.

Need to fix up xmpp info plugin to use the correct thread as well.

Also, right now tweets indiscriminately when you love, should turn that
off before merging to master.