* Checking the reply status during the last.fm authentication process
The last.fm API returns "ok" on the status attribute of the reply if
the authentication succeeded and "failed" otherwise [1]. So this patch
verifies if the status returned is "ok" to proceed with the
authentication.

This patch also improves the message displayed in the log when an
authentication error occurs. Now the error message returned by the
server is printed out even when the QNetworkReply::error() != NoError
(the message retuned by the server is a lot more informative).

[1] - http://www.lastfm.it/api/show/auth.getMobileSession