Mark functions that should only be used internally.
JSResolverHelper slots and INVOKABLEs are exported to JavaScript but not
all functions shall be used a resolver. Some of the exported functions
are only meant to be used by Wrappers in tomahawk.js as they cannot
fulfil their task completely in JS and need to callback to Qt for
certain functionality.