Don't pass an Url needlessly through QUrl
This does more harm than good: If the first part after the protocol is
numeric QUrl assumes this is a byteencoded IPv4 address and converts it
to the dot notation. If we actually want to pass a numeric ID here, this
will turn the ID into an IP which is then useless for us.