Install snoresettings as "Notification Settings".
This enables the user to change the notification backend
and to configure the banckend.
Opening the shortcut will only work after Tomahawk was started
for a first time to initialize the settings.
An alternative would be to call snoresettings.exe from inside of
Tomahawk or to use the QWidget provided by libsnore.